Better way to include core.sh, more arguments for scramdir.sh

+++ scramdir.sh
@@ -1,30 +1,71 @@
 #!/bin/bash
 
-source core.sh
+source $(dirname $(readlink -f $0))/core.sh
+
+TRUE=1
+FALSE=0
 
 help() {
 	phelp $0
 }
 
-BASE_FOLDER="$1"
-BASE_FOLDER=${BASE_FOLDER:-.}
-BASE_FOLDER=${BASE_FOLDER%/}
-FILE_LENGTH=7
-
 get_random_name() {
 	cat /dev/urandom | tr -cd 'a-z0-9' | head -c $FILE_LENGTH
 }
 
+FILE_LENGTH=7
+
+force=$FALSE
+quiet=$FALSE
+simulate=$FALSE
+
+# Read parameters
+while test $# -gt 0; do
+	case "$1" in
+		-h|--help)
+			help
+			exit 0
+			;;
+		-f|--force)
+			force=$TRUE
+			shift
+			;;
+		-q|--quiet)
+			quiet=$TRUE
+			shift
+			;;
+		-s|--simulate)
+			simulate=$TRUE
+			shift
+			;;
+		*)
+			if [ ! -z "$BASE_FOLDER" ]; then
+				perror "Base folder already specified"
+				exit 1
+			fi
+
+			BASE_FOLDER="$1"
+			BASE_FOLDER=${BASE_FOLDER%/}
+			shift
+			;;
+	esac
+done
+
+# Default value of the current directory
+BASE_FOLDER=${BASE_FOLDER:-.}
+
 if [ ! -d "$BASE_FOLDER" ]; then
 	echo "Directory \"$BASE_FOLDER\" does not exist"
-	exit 1
+	exit 2
 fi
 
-# Ask for confirmation
-code=$(confirm "Really shuffle all files in $(readlink -f $BASE_FOLDER)? [y/n] ")
+if [ $force -eq "$FALSE" ]; then
+	# Ask for confirmation
+	code=$(confirm "Really shuffle all files in $(readlink -f $BASE_FOLDER)? [y/n] ")
 
-if [ $code -ne 0 ]; then
-	exit 1
+	if [ $code -ne 0 ]; then
+		exit 3
+	fi
 fi
 
 counter=0
@@ -48,9 +89,14 @@ for f in $BASE_FOLDER/*; do
 
 	random=$(get_random_name)
 	full_path="$BASE_FOLDER/$random.$extension"
-	echo "$f --> $full_path" 1>&2
-	mv "$f" "$full_path"
-	counter=$[counter + 1]
+	if [ "$quiet" -eq "$FALSE" ]; then
+		echo "$f --> $full_path"
+	fi
+
+	if [ "$simulate" -eq "$FALSE" ]; then
+		mv "$f" "$full_path"
+	fi
+	counter=$((counter + 1))
 done
 
-echo "Renamed $counter files"
+echo "Renamed $counter files in $(readlink -f $BASE_FOLDER)"
